requirements.txt
piexif
pillow
fire
tqdm
geoclip
geopy